pyproject.toml
[tool.poetry]
name = "pw-homework-module-13"
version = "0.1.0"
description = ""
authors = ["RuslanaPashkutska <ryslana.pashkytska@gmail.com>"]
packages = [{ include = "src" }]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
alembic = ">=1.16.2,<2.0.0"
dotenv = ">=0.9.9,<0.10.0"
pydantic-settings = ">=2.10.1,<3.0.0"
psycopg2-binary = ">=2.9.10,<3.0.0"
fastapi = ">=0.110.0,<1.0.0"
uvicorn = { version = ">=0.29.0,<1.0.0", extras = ["standard"] }
python-jose = { version = ">=3.3.0,<4.0.0", extras = ["cryptography"] }
passlib = { version = ">=1.7.4,<2.0.0", extras = ["bcrypt"] }
cloudinary = ">=1.37.0,<2.0.0"
redis = "^6.2.0"
fastapi-mail = "^1.5.0"
fastapi-limiter = "^0.1.6"
django-environ = "^0.12.0"
django = "^5.2.4"
asyncpg = "^0.30.0"
psycopg = {extras = ["binary"], version = "^3.2.9"}
python-multipart = "^0.0.20"
greenlet = "^3.2.3"
sphinx = "^8.2.3"
httpx = "^0.28.1"
pytest = "^8.4.1"
aiosqlite = "^0.21.0"
pytest-asyncio = "^1.0.0"

[tool.poetry.group.dev.dependencies]
sphinx = "^8.2.3"
pytest = "^8.4.1"
pytest-cov = "^6.2.1"
pytest-asyncio = "^1.0.0"
bcrypt = "<4"
setuptools = "^80.9.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
addopts = "--cov=src --cov-report=html --cov-report=term-missing"
testpaths = "tests"

[tool.coverage.run]
source = ["src"]
branch = true
omit = [
    "*/__init__.py",
    "*/__main__.py",

]

[tool.coverage.report]
show_missing = true
exclude_lines = [
    "if TYPE_CHECKING:",
    "@(abc\\.)?abstractmethod",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "pass",
]